Multi tags and integrations + permissions (#165)
* Making a real test plugin.

* working on adding an overload decorator

* trying to figure out the doc string

* Overloads

* giving up on overload

* finished things up

* reduced scope

* added mkdir and cp

* Updated to the newer enum

* commented old stuff

* reverted the old test file

* working more on the plugin

* changed enum to avoid collisions

* fixed some problems with the fileexplorer

* fixed issues with the untested api. Wrote a test for it but it requires a file

* deleting overload

* deleted unuse imports

* deleted unused imports and useless serializer

* added types for the callbacks

* added permissions into plugins

* Change category to tag array and permissions to permission array. Add integration array

* Add missing code to api's plugin, and change dictionary of string to array of hashed values

* Split EIntegration and EIntegrationRequest

* Fix integration hash usage and warning

* Readd missing export locations and add testing to UIPlugin

+++ nanome/api/files.py
@@ -12,6 +12,8 @@ class Files(_Files):
         :param callback: function that will be called with the full filename
         :type callback: method (:class:`~nanome.util.file.FileError`, str) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
         expects_response = callback is not None or self.plugin.is_async
         id = self.plugin._network._send(_Messages.print_working_directory, None, expects_response)
         return self.plugin._save_callback(id, callback)
@@ -24,6 +26,8 @@ class Files(_Files):
         :param callback: called when operation is completed with any potential errors.
         :type callback: method (:class:`~nanome.util.file.FileError`) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
         expects_response = callback is not None or self.plugin.is_async
         id = self.plugin._network._send(_Messages.cd, directory, expects_response)
         return self.plugin._save_callback(id, callback)
@@ -36,6 +40,8 @@ class Files(_Files):
         :param callback: function that will be called with contents of the directory
         :type callback: method (:class:`~nanome.util.file.FileError`, list of :class:`~nanome.util.file.FileMeta`) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
         expects_response = callback is not None or self.plugin.is_async
         id = self.plugin._network._send(_Messages.ls, directory, expects_response)
         return self.plugin._save_callback(id, callback)
@@ -50,6 +56,8 @@ class Files(_Files):
         :param callback: called when operation is completed with any potential errors.
         :type callback: method (:class:`~nanome.util.file.FileError`) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
         expects_response = callback is not None or self.plugin.is_async
         id = self.plugin._network._send(_Messages.mv, (source, dest), expects_response)
         return self.plugin._save_callback(id, callback)
@@ -64,6 +72,9 @@ class Files(_Files):
         :param callback: called with the path to the file and any potential errors
         :type callback: method (:class:`~nanome.util.file.FileError`, str) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
+
         def cb(error, file):
             if error == FileErrorCode.no_error:
                 with open(dest, 'wb') as ofile:
@@ -88,6 +99,8 @@ class Files(_Files):
         :param callback: called when operation is completed with any potential errors.
         :type callback: method (:class:`~nanome.util.file.FileError`) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
         with open(source, "rb") as f:
             file = f.read()
             f.close()
@@ -103,6 +116,8 @@ class Files(_Files):
         :param callback: called when operation is completed with any potential errors.
         :type callback: method (:class:`~nanome.util.file.FileError`) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
         expects_response = callback is not None or self.plugin.is_async
         id = self.plugin._network._send(_Messages.rm, target, expects_response)
         return self.plugin._save_callback(id, callback)
@@ -115,6 +130,8 @@ class Files(_Files):
         :param callback: called when operation is completed with any potential errors.
         :type callback: method (:class:`~nanome.util.file.FileError`) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
         expects_response = callback is not None or self.plugin.is_async
         id = self.plugin._network._send(_Messages.rmdir, target, expects_response)
         return self.plugin._save_callback(id, callback)
@@ -129,6 +146,8 @@ class Files(_Files):
         :param callback: called when operation is completed with any potential errors.
         :type callback: method (:class:`~nanome.util.file.FileError`) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
         expects_response = callback is not None or self.plugin.is_async
         id = self.plugin._network._send(_Messages.cp, (source, dest), expects_response)
         return self.plugin._save_callback(id, callback)
@@ -141,6 +160,8 @@ class Files(_Files):
         :param callback: called when operation is completed with any potential errors.
         :type callback: method (:class:`~nanome.util.file.FileError`) -> None
         """
+        if self.plugin._permissions.get('files', 0 ) <= 0:
+            raise Exception("Plugin requires files permission to use this method.")
         expects_response = callback is not None or self.plugin.is_async
         id = self.plugin._network._send(_Messages.mkdir, target, expects_response)
         return self.plugin._save_callback(id, callback)
